update check in code
