+++ functions/src/apiUtils.ts
@@ -1,3 +1,5 @@
+import { IGetUserAuthInfoRequest, IProfile } from './types';
+import * as admin from 'firebase-admin';
 type HangzoneApiErrorStatus = 403 | 404 | 422 | 500;
 type ResourceDict = { [resourceName: string]: any };
 
@@ -23,4 +25,18 @@ export function requiredParameterChecker(params: ResourceDict): string | null {
     if (!val) return `${key} is a required parameter`;
   }
   return null;
+}
+
+export async function getOwnProfile(req: IGetUserAuthInfoRequest) {
+  const firebaseUserUID = req.user?.uid;
+  if (!firebaseUserUID) throw new Error('Not a valid user');
+
+  const profile = (await admin
+    .firestore()
+    .collection('profiles')
+    .doc(firebaseUserUID)
+    .get()
+    .then((doc) => doc.data())) as IProfile | null;
+
+  return profile;
 }
